requirements.txt
earthengine-api>=0.1.370
geopandas>=0.14.0
numpy>=1.26.0
Pillow>=10.0.0
pycairo>=1.25.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
rasterio>=1.3.0
shapely>=2.0.0
opencv-python>=4.9.0
imageio[ffmpeg]>=2.34.0
